atom lens and derived atom
Derived atoms, read-write and read-only, added

Useful, even if they don't work with the async or soliton.sm specific
functionality.

TODO? Maybe have focus/over/put passed in? With async fns for over ... does
swap! return a channel then, that's super weird and definitely not a good idea

If there are future changes to soliton.sm that allow forridan-style step fn
swapping mid-pipeline then a properly acting atom lens could be made. When it
arrives at the atom lens, it takes the rest of the compound lens and applies
atomically using the swap! fn to the atom, basically splitting the put/over
operation into outside the atom and inside the atom.